Inline Standalone within Preview
Training-preview being removed, Standalone app is not required anymore
This patch is integrating Standalone within Preview and remove the
Standalone app
