+++ preview/app/controllers/FaviconController.scala
@@ -5,7 +5,7 @@ import play.api.mvc.Controller
 /*
  * Pointless controller, only here to work around a bug in the Play routes compiler.
 
- * If you try to point directly to controller.Assets.at(...) in standalone.routes
+ * If you try to point directly to controller.Assets.at(...) in routes
  * then you will get strange compilation errors like:
  * - patterns after a variable pattern cannot match (SLS 8.1.1)
  * - unreachable code due to variable pattern 'file'
