+++ preview/app/controllers/OAuthLoginPreviewController.scala
@@ -1,17 +1,17 @@
 package controllers
 
 import com.gu.googleauth.{GoogleAuthConfig, UserIdentity}
-import googleAuth.OAuthLoginController
-import play.api.mvc.{Action, AnyContent, Request}
 import conf.Configuration
+import googleAuth.OAuthLoginController
 import model.ApplicationContext
 import play.api.libs.crypto.CryptoConfig
 import play.api.libs.ws.WSClient
+import play.api.mvc.{Action, AnyContent, Request}
 
-class OAuthLoginStandaloneController(val wsClient: WSClient, val cryptoConfig: CryptoConfig)(implicit context: ApplicationContext) extends OAuthLoginController {
+class OAuthLoginPreviewController(val wsClient: WSClient, val cryptoConfig: CryptoConfig)(implicit context: ApplicationContext) extends OAuthLoginController {
 
   override def login = Action { request =>
-    Ok(views.html.standalone_auth(context.applicationIdentity.name, "Dev", UserIdentity.fromRequest(request)))
+    Ok(views.html.previewAuth(context.applicationIdentity.name, "Dev", UserIdentity.fromRequest(request)))
   }
   override def googleAuthConfig(request: Request[AnyContent]): Option[GoogleAuthConfig] = Configuration.standalone.oauthCredentials.map { cred =>
     GoogleAuthConfig(
